+++ www/src/components/layout.js
@@ -148,9 +148,13 @@ class DefaultLayout extends React.Component {
     return (
       <div className={isHomepage ? `is-homepage` : ``}>
         <Helmet defaultTitle={`GatsbyJS`} titleTemplate={`%s | GatsbyJS`}>
+          <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1, shrink-to-fit=no, viewport-fit=cover"
+          />
           <meta name="twitter:site" content="@gatsbyjs" />
-          <meta name="og:type" content="website" />
-          <meta name="og:site_name" content="GatsbyJS" />
+          <meta property="og:type" content="website" />
+          <meta property="og:site_name" content="GatsbyJS" />
           <meta name="docsearch:version" content="2.0" />
           <link
             rel="canonical"
@@ -194,6 +198,8 @@ class DefaultLayout extends React.Component {
                 ? presets.bannerHeight
                 : `calc(${presets.bannerHeight} + ${presets.headerHeight})`,
             },
+            paddingLeft: `env(safe-area-inset-left)`,
+            paddingRight: `env(safe-area-inset-right)`,
           }}
         >
           <PageWithSidebar